requirements/dev.txt
aspy.yaml==1.3.0          # via pre-commit
astor==0.8.0              # via wemake-python-styleguide
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via flake8-bugbear, flake8-eradicate, packaging, pytest, wemake-python-styleguide
bandit==1.6.2             # via flake8-bandit
cfgv==2.0.1               # via pre-commit
click==7.0                # via flask, pip-tools
colorama==0.3.9           # via radon
coverage==4.5.4           # via pytest-cov
docutils==0.15.2          # via restructuredtext-lint
entrypoints==0.3          # via flake8
environs==5.2.1
eradicate==1.0            # via flake8-eradicate
flake8-annotations-complexity==0.0.2  # via wemake-python-styleguide
flake8-bandit==2.1.1      # via wemake-python-styleguide
flake8-broken-line==0.1.1  # via wemake-python-styleguide
flake8-bugbear==19.8.0    # via wemake-python-styleguide
flake8-builtins==1.4.1    # via wemake-python-styleguide
flake8-coding==1.3.2      # via wemake-python-styleguide
flake8-commas==2.0.0      # via wemake-python-styleguide
flake8-comprehensions==2.2.0  # via wemake-python-styleguide
flake8-debugger==3.1.0    # via wemake-python-styleguide
flake8-docstrings==1.4.0  # via wemake-python-styleguide
flake8-eradicate==0.2.1   # via wemake-python-styleguide
flake8-executable==2.0.3  # via wemake-python-styleguide
flake8-isort==2.7.0       # via wemake-python-styleguide
flake8-logging-format==0.6.0  # via wemake-python-styleguide
flake8-pep3101==1.2.1     # via wemake-python-styleguide
flake8-polyfill==1.0.2    # via flake8-bandit, pep8-naming, radon
flake8-print==3.1.0       # via wemake-python-styleguide
flake8-quotes==2.1.0      # via wemake-python-styleguide
flake8-rst-docstrings==0.0.11  # via wemake-python-styleguide
flake8-string-format==0.2.3  # via wemake-python-styleguide
flake8==3.7.8             # via flake8-bandit, flake8-broken-line, flake8-bugbear, flake8-builtins, flake8-coding, flake8-commas, flake8-comprehensions, flake8-debugger, flake8-docstrings, flake8-eradicate, flake8-executable, flake8-isort, flake8-pep3101, flake8-polyfill, flake8-print, flake8-quotes, flake8-rst-docstrings, flake8-string-format, wemake-python-styleguide, yesqa
flask==1.1.1
gitdb2==2.0.5             # via gitpython
gitpython==3.0.2          # via bandit
identify==1.4.7           # via pre-commit
importlib-metadata==0.23  # via pluggy, pre-commit, pytest
importlib-resources==1.0.2  # via pre-commit
isort==4.3.21             # via flake8-isort
itsdangerous==1.1.0       # via flask
jinja2==2.10.1            # via flask
mando==0.6.4              # via radon
markupsafe==1.1.1         # via jinja2
marshmallow==3.1.1        # via environs
mccabe==0.6.1             # via flake8
more-itertools==7.2.0     # via pytest, zipp
mypy-extensions==0.4.1    # via mypy
mypy==0.720
nodeenv==1.3.3            # via pre-commit
packaging==19.1           # via pytest
pathspec==0.5.9           # via yamllint
pbr==5.4.3                # via stevedore
pep8-naming==0.8.2        # via wemake-python-styleguide
pip-tools==4.1.0
pluggy==0.13.0            # via pytest
pre-commit==1.18.3
py==1.8.0                 # via pytest
pycodestyle==2.5.0        # via flake8, flake8-bandit, flake8-debugger, flake8-print
pydocstyle==4.0.1         # via flake8-docstrings
pyflakes==2.1.1           # via flake8
pygments==2.4.2
pyhamcrest==1.9.0
pyparsing==2.4.2          # via packaging
pytest-cov==2.7.1
pytest==5.1.2
python-dotenv==0.10.3     # via environs
pyyaml==5.1.2             # via aspy.yaml, bandit, pre-commit, yamllint
radon==2.4.0              # via wemake-python-styleguide
restructuredtext-lint==1.3.0
six==1.12.0               # via bandit, cfgv, flake8-print, mando, packaging, pip-tools, pre-commit, pyhamcrest, stevedore
smmap2==2.0.5             # via gitdb2
snowballstemmer==1.9.1    # via pydocstyle
stevedore==1.31.0         # via bandit
testfixtures==6.10.0      # via flake8-isort
tokenize-rt==3.2.0        # via yesqa
toml==0.10.0              # via pre-commit
typed-ast==1.4.0          # via mypy
typing-extensions==3.7.4  # via mypy, wemake-python-styleguide
virtualenv==16.7.5        # via pre-commit
wcwidth==0.1.7            # via pytest
wemake-python-styleguide==0.12.4
werkzeug==0.15.6          # via flask
yamllint==1.17.0
yesqa==0.0.11
zipp==0.6.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.2.0        # via flake8-annotations-complexity, pyhamcrest
